move server actions to auth service file
